Only show the agenda part of the journal entry.
For some reason, possibly related to Babel and/or ES6, ESLint's
`no-unused-vars` rule is failing to ignore the `_` variable I use to
pull the toplevel elements out of the intermediate parsed markup.

So I configured that to be a warning instead of an error, but
Overcommit is ignoring my configuration, so I uninstalled it.